*Formats Beta*
Added RF64 EBU Tech 3306

RiffReader - Updated for 64bit support (BWAV), Node which are read now contain the SubType to reduce IO.

BaseMediaReader - 

Corrected logic in ReadNext for reading large boxes when the type is UUID, stored UUID in node Identifier to reduce IO.

Corrected logic in ReadBoxes where count was not being correctly calculated resulting in more boxes being returned than desired.

Sets Volume on created Tracks.

Ensured SampleTables cannot overflow parsing.

Optomized GetTracks implementation, still WIP.

MediaFileStream - ReadBytes now accepts long.

ASFReader - Fixed double seeking position when reading objects.

Track - Now has Volume property.

+++ Containers/MediaFileStream.cs
@@ -502,12 +502,12 @@ namespace Media.Container
         /// <param name="count"></param>
         /// <returns></returns>
         [System.Runtime.CompilerServices.MethodImpl(System.Runtime.CompilerServices.MethodImplOptions.Synchronized)]
-        public virtual byte[] ReadBytes(int count)
+        public virtual byte[] ReadBytes(long count)
         {
             if (count <= 0) return Media.Common.MemorySegment.EmptyBytes; 
             byte[] result = new byte[count]; 
-            int i = 0; 
-            /*do*/while ((count -= (i += Read(result, i, count))) > 0) ; 
+            long i = 0; 
+            /*do*/while ((count -= (i += Read(result, (int)i, (int)count))) > 0) ; 
             return result;
         }
